self_pruning_nn/requirements.txt
torch
torchvision
pytest
numpy
pyyaml
matplotlib
pandas
tqdm
fastapi
uvicorn
python-multipart
